# chore: common linq loop allocations, add small optimisations 