--- a/src/Riok.Mapperly/Helpers/NullableSymbolExtensions.cs
+++ b/src/Riok.Mapperly/Helpers/NullableSymbolExtensions.cs
@@ -5,8 +5,6 @@ namespace Riok.Mapperly.Helpers;
 
 public static class NullableSymbolExtensions
 {
-    private const string NullableGenericTypeName = "System.Nullable<T>";
-
     internal static bool HasSameOrStricterNullability(this ITypeSymbol symbol, ITypeSymbol other)
     {
         return symbol.NullableAnnotation == NullableAnnotation.NotAnnotated
@@ -109,10 +107,7 @@ public static class NullableSymbolExtensions
 
     private static ITypeSymbol? NonNullableValueType(this ITypeSymbol symbol)
     {
-        if (
-            symbol is INamedTypeSymbol { IsValueType: true, IsGenericType: true } namedType
-            && namedType.ConstructedFrom.ToDisplayString() == NullableGenericTypeName
-        )
+        if (symbol.IsValueType && symbol is INamedTypeSymbol { OriginalDefinition.SpecialType: SpecialType.System_Nullable_T } namedType)
             return namedType.TypeArguments[0];
         return null;
     }