# Remove linq allocations, add small optimisations

## Description
Slightly microbenchmarky improvements, not sure about the  manual foreach loops

- Remove `FirstOrDefault` allocation from `SymbolExtensions.ImplementsGeneric` - 106KB - `Func<INamedTypeSymbol, Boolean>`
- Remove `Where` allocation from `SymbolAccessor.GetMappableMembers` - 76KB - `Func<ISymbol, Boolean>`
- Remove `Where` allocation from `SymbolAccessor.GetAttributes` - 150KB - `Func<AttributeData, Boolean>`
- Use non string comparison `NonNullableValueType` check
- Add `IsRefLike` check to `GetEnumeratedType`
- Remove unneeded `types.Get<int>()` call  from `CollectionInfoBuilder.FindCountProperty`


<!-- Please include a summary of the changes and the related issue. -->

Related to #530

## Benchmarks
200 KB saved, little change in speed.
|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.441 ms | 0.0206 ms | 0.0237 ms |   78.1250 |  15.6250 |  162.41 KB |
| LargeCompile | 56.606 ms | 1.1015 ms | 1.4323 ms | 1428.5714 | 428.5714 | 9367.77 KB |

### Original
|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.450 ms | 0.0288 ms | 0.0374 ms |   78.1250 |  19.5313 |  167.37 KB |
| LargeCompile | 55.899 ms | 1.0530 ms | 1.3318 ms | 1428.5714 | 428.5714 | 9579.64 KB |

### Changes

## Checklist

- [x] The existing code style is followed
- [x] The commit message follows our guidelines
- [x] Performed a self-review of my code
- [x] Hard-to-understand areas of my code are commented